pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lunarpits"
version = "0.1.0"
description = "Lunar pit imagery processing, indexing, tiling, annotation, feature, and model tooling."
requires-python = ">=3.11"

[project.optional-dependencies]
qa = ["matplotlib", "numpy", "rasterio"]
location = ["matplotlib", "numpy", "pandas", "pyarrow", "rasterio", "PyYAML"]
tiles = ["numpy", "pandas", "geopandas", "shapely", "rasterio", "pyproj", "requests", "tqdm", "pytest"]

[project.scripts]
lunarpits-process-lroc = "lunarpits.processing.lroc_product:main"
lunarpits-qa-lroc-tif = "lunarpits.processing.raster_qa:main"
lunar-tiles = "lunar_tile_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
